server/requirements.txt
paho-mqtt
numpy
pyaudio